refactor: 공유 카드 섀도우 추가, 코드 분리
